Switch to remix-express-vite-plugin
