+++ package.json
@@ -4,8 +4,8 @@
   "sideEffects": false,
   "type": "module",
   "scripts": {
+    "dev": "cross-env NODE_ENV=development SECRETS_OVERRIDE=1 vite",
     "build": "remix vite:build",
-    "dev": "cross-env NODE_ENV=development SECRETS_OVERRIDE=1 remix dev -c \"node --inspect ./build/server/index.js\"",
     "docker": "docker compose up -d",
     "lint": "oxlint && eslint",
     "db:generate": "prisma generate",
@@ -62,6 +62,7 @@
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
     "react-use": "^17.6.0",
+    "remix-create-express-app": "^0.4.5",
     "remix-typedjson": "^0.4.1",
     "samlify": "2.9.1",
     "validator": "^13.15.0",
@@ -104,6 +105,7 @@
     "oxlint": "^0.16.4",
     "prettier": "3.5.3",
     "prisma": "^6.5.0",
+    "remix-express-dev-server": "^0.4.4",
     "start-server-and-test": "^2.0.11",
     "tsconfig-paths": "^4.2.0",
     "tsx": "^4.19.3",
